typescripted most everything, need to squash a lot of new bugs and fix Button behavior after removing 'new' keyword to appease typescript
