+++ gulpfile.js
@@ -1,0 +1,42 @@
+var gulp = require("gulp");
+var browserify = require("browserify");
+var source = require("vinyl-source-stream");
+var tsify = require("tsify");
+var ts = require("gulp-typescript");
+var tsProject = ts.createProject("tsconfig.json");
+gulp.task("copy-scripts", function () {
+  return tsProject.src().pipe(tsProject()).js.pipe(gulp.dest("dist"));
+});
+gulp.task('copy-index', function() {
+    return gulp.src('./src/index.html').pipe(gulp.dest('./dist'));
+});
+gulp.task('copy-favicon', function() {
+    return gulp.src('./src/favicon.ico').pipe(gulp.dest('./dist'));
+});
+gulp.task('copy-css', function() {
+    return gulp.src('./src/css/*.css').pipe(gulp.dest('./dist/css'));
+});
+gulp.task('copy-libs', function() {
+    return gulp.src('./src/lib/*.js').pipe(gulp.dest('./dist/lib'));
+});
+gulp.task('browserify', function() {
+    return browserify({
+        basedir: ".",
+        debug: true,
+        entries: ["src/script/engine.ts"],
+        cache: {},
+        packageCache: {},
+      })
+        .plugin(tsify)
+        .bundle()
+        .pipe(source("bundle.js"))
+        .pipe(gulp.dest("dist"));
+});
+gulp.task('default', gulp.series(
+    // 'copy-scripts',
+    'copy-index',
+    // 'copy-favicon',
+    // 'copy-css',
+    // 'copy-libs',
+    'browserify'
+));
