+++ src/index.html
@@ -22,11 +22,11 @@
 			document.write('<script src="lib/jquery.min.js"><\/script>')
 		}
 	</script>
-	<script src="lib/jquery.color-2.1.2.min.js"></script>
+	<!-- <script src="lib/jquery.color-2.1.2.min.js"></script>
 	<script src="lib/jquery.event.move.js"></script>
 	<script src="lib/jquery.event.swipe.js"></script>
 	<script src="lib/base64.js"></script>
-	<script src="lib/translate.js"></script>
+	<script src="lib/translate.js"></script> -->
 	
 	<script>
 		// try to read "lang" param's from url
@@ -44,7 +44,9 @@
 		}
 	</script>
 	
-	<script src="script/Button.js"></script>
+	<script src="bundle.js"></script>
+	
+	<!-- <script src="script/Button.js"></script>
 	<script src="script/state_manager.js"></script>
 	<script src="script/player/itemList.js"></script>
 	<script src="script/header.js"></script>
@@ -57,11 +59,11 @@
 	<script src="script/weather.js"></script>
 	<script src="script/places/room.js"></script>
 	<script src="script/places/road.js"></script>
-	<script src="script/places/outpost.js"></script>
+	<script src="script/places/outpost.js"></script> -->
 	<!-- Event modules -->
-	<script src="script/events/room.js"></script>
+	<!-- <script src="script/events/room.js"></script>
 	<script src="script/events/roadWander.js"></script>
-	<script src="script/engine.js"></script>
+	<script src="script/engine.js"></script> -->
 
 	
 	<script type='text/javascript'>
@@ -80,7 +82,9 @@
 	<div class="rain front-row"></div>
   	<div class="rain back-row"></div>	
 	<div id="wrapper">
-		<div id="saveNotify"><script>document.write(_("saved."));</script></div>
+		<div id="saveNotify">
+			<!-- <script>document.write(_("saved."));</script> -->
+		</div>
 		<div id="content">
 			<div id="outerSlider">
 				<div id="main">
