+++ src/lib/base64.js
