+++ src/lib/jquery.event.move.js
