+++ src/lib/jquery.event.swipe.js
